myutils_dpy/pyproject.toml
[tool.poetry]
name = "myutils_dpy"
version = "0.1.0"
description = ""
authors = ["owner <yokoyamacode@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
